Split MockStorageReference in own file and add many mock methods to it.
This commit will put `MockStorageReference` in his own file; other than that it implements mocks for many get methods like `getPath` and `getStorage`.
`MockFirebaseStorage` now contains two maps: one with all the stored files, the other with all the stored Uint8List data; calls to `putFile`, `putData` and `delete` will write to that lists.
`MockStorageUploadTask` and `MockStorageTaskSnapshot` now takes an instance of `StorageReference` in order to return it in `ref` method.